requirements.txt
accelerate==0.24.1
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.3
asyncer==0.0.5
attrs==23.2.0
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cmake==3.25.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
dearpygui==1.7.1
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
Deprecated==1.2.14
diff-gaussian-rasterization @ file:///mnt/HDD2/khanh/content/projects/dreamgaussian/diff-gaussian-rasterization
diffusers==0.24.0
einops==0.4.1
exceptiongroup @ file:///croot/exceptiongroup_1706031385326/work
executing==2.0.1
fastapi==0.110.0
ffmpy==0.3.2
filelock @ file:///croot/filelock_1700591183607/work
filetype==1.2.0
flatbuffers==24.3.7
fonttools==4.49.0
frozenlist==1.4.1
fsspec @ file:///croot/fsspec_1701286474621/work
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
gradio==4.8.0
gradio_client==0.7.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.19.4
humanfriendly==10.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.21.1
imageio-ffmpeg==0.4.8
importlib_metadata==7.0.2
importlib_resources==6.1.3
ipython @ file:///croot/ipython_1704833016303/work
jaxtyping==0.2.28
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiui @ git+https://github.com/ashawkey/kiuikit@a69f0de046e39ea5a57f19695fdfdb84318607ec
kiwisolver==1.4.5
lazy_loader==0.3
lightning==2.2.1
lightning-utilities==0.11.0
lit==15.0.7
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
marshmallow==3.21.1
matplotlib==3.8.0
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.5
mypy-extensions==1.0.0
networkx @ file:///croot/networkx_1690561992265/work
ninja==1.10.2.3
numba==0.59.0
numpy==1.22.4
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast/@c5caf7bdb8a2448acc491a9faa47753972edd380
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
objprint==0.2.3
omegaconf==2.3.0
onnxruntime==1.17.1
onnxruntime-gpu==1.16.3
opencv-python==4.8.0.76
opencv-python-headless==4.9.0.80
orjson==3.9.15
packaging==24.0
pandas==1.4.2
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow @ file:///croot/pillow_1707233021655/work
platformdirs==4.2.0
plyfile==0.7.4
pooch==1.8.1
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==4.25.3
psutil==5.9.8
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
pygltflib==1.16.0
Pygments @ file:///croot/pygments_1684279966437/work
PyMatting==1.1.12
PyMCubes==0.1.2
pymeshlab==2022.2.post3
pyparsing==3.1.2
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-lightning==2.2.1
pytz==2024.1
PyWavefront==1.3.3
PyWavelets==1.5.0
PyYAML @ file:///croot/pyyaml_1698096049011/work
referencing==0.33.0
regex==2023.12.25
rembg==2.0.49
requests @ file:///croot/requests_1707355572290/work
rich==13.4.2
rpds-py==0.18.0
safetensors==0.4.2
scikit-image==0.20.0
scikit-learn==1.0.2
scipy==1.11.4
semantic-version==2.10.0
shellingham==1.5.4
simple-knn @ file:///mnt/HDD2/khanh/content/projects/dreamgaussian/simple-knn
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.36.3
sympy @ file:///croot/sympy_1701397643339/work
threadpoolctl==3.3.0
tifffile==2024.2.12
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchmetrics==1.3.2
torchvision==0.15.2+cu118
tqdm==4.65.2
traitlets @ file:///croot/traitlets_1671143879854/work
transformers==4.36.2
trimesh==4.0.5
triton==2.0.0
typeguard==2.13.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3 @ file:///croot/urllib3_1707770551213/work
uvicorn==0.28.0
varname==0.13.0
watchdog==4.0.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websockets==11.0.3
wrapt==1.16.0
xatlas==0.0.6
yarl==1.9.4
zipp==3.17.0
